Initial commit
I was learning about standard C's signal and raise functions and apparently
they're not very portable and have lots of undefined/implementation defined
behavior.  So I figured I'd see how I'd make something similar myself.  This
is just a start and I'm not if/when I'll use this but it is interesting and
one way to handle certain kinds of C errors.