1.0 Basically done.  It is a simple idea
I only barely test it in main but it's so simple.  When I try using it
for some project it will get much more coverage and experimentation with
how it's best used.  I have some ideas but I need to pick a project or
make a branch in CVector or something.